+++ qutip/operators.py
@@ -127,8 +127,8 @@ def _jplus(j):
     m = np.arange(j, -j - 1, -1, dtype=complex)
     data = (np.sqrt(j * (j + 1.0) - (m + 1.0) * m))[1:]
     N = m.shape[0]
-    ind = np.arange(1, N, dtype=np.int32)
-    ptr = np.array(list(range(N-1))+[N-1]*2, dtype=np.int32)
+    ind = np.arange(1, N, dtype=np.int64)
+    ptr = np.array(list(range(N-1))+[N-1]*2, dtype=np.int64)
     ptr[-1] = N-1
     return fast_csr_matrix((data,ind,ptr), shape=(N,N))
 
@@ -141,14 +141,14 @@ def _jz(j):
     data = np.array([j-k for k in range(N) if (j-k)!=0], dtype=complex)
     # Even shaped matrix
     if (N % 2 == 0):
-        ind = np.arange(N, dtype=np.int32)
-        ptr = np.arange(N+1,dtype=np.int32)
+        ind = np.arange(N, dtype=np.int64)
+        ptr = np.arange(N+1,dtype=np.int64)
         ptr[-1] = N
     # Odd shaped matrix
     else:
         j = int(j)
-        ind = np.array(list(range(j))+list(range(j+1,N)), dtype=np.int32)
-        ptr = np.array(list(range(j+1))+list(range(j,N)), dtype=np.int32)
+        ind = np.array(list(range(j))+list(range(j+1,N)), dtype=np.int64)
+        ptr = np.array(list(range(j+1))+list(range(j,N)), dtype=np.int64)
         ptr[-1] = N-1
     return fast_csr_matrix((data,ind,ptr), shape=(N,N))
 
@@ -377,8 +377,8 @@ shape = [4, 4], type = oper, isHerm = False
     if not isinstance(N, (int, np.integer)):  # raise error if N not integer
         raise ValueError("Hilbert space dimension must be integer value")
     data = np.sqrt(np.arange(offset+1, N+offset, dtype=complex))
-    ind = np.arange(1,N, dtype=np.int32)
-    ptr = np.arange(N+1, dtype=np.int32)
+    ind = np.arange(1,N, dtype=np.int64)
+    ptr = np.arange(N+1, dtype=np.int64)
     ptr[-1] = N-1
     return Qobj(fast_csr_matrix((data,ind,ptr),shape=(N,N)), isherm=False)
 
@@ -554,13 +554,13 @@ shape = [4, 4], type = oper, isHerm = True
     """
     if offset == 0:
         data = np.arange(1,N, dtype=complex)
-        ind = np.arange(1,N, dtype=np.int32)
-        ptr = np.array([0]+list(range(0,N)), dtype=np.int32)
+        ind = np.arange(1,N, dtype=np.int64)
+        ptr = np.array([0]+list(range(0,N)), dtype=np.int64)
         ptr[-1] = N-1
     else:
         data = np.arange(offset, offset + N, dtype=complex)
-        ind = np.arange(N, dtype=np.int32)
-        ptr = np.arange(N+1,dtype=np.int32)
+        ind = np.arange(N, dtype=np.int64)
+        ptr = np.arange(N+1,dtype=np.int64)
         ptr[-1] = N
 
     return Qobj(fast_csr_matrix((data,ind,ptr), shape=(N,N)), isherm=True)
